Remove metadata of reference tables after master_remove_node
With this change, we start to delete placement of reference tables at given worker node
after master_remove_node UDF call. We remove placement metadata at master node but we do
not drop actual shard from the worker node. Because to support transactional
master_remove_node, we have to use coordinated transaction so that if master ROLLBACKs
we can ROLLBACK at the worker too. However to be able to use coordinated transactions, we
need to keep the connection open after removing the node from cluster, which may cause
leaking connections. Therefore we prefered to not dropping actual shard. In case of
re-adding the node to the cluster, master_add_node will DROP and re-create all reference
tables at that node.